move rl part from `torchkit` to `rl`
