+++ policies/models/policy_rnn_shared.py
@@ -8,9 +8,9 @@ import torch.nn as nn
 from torch.nn import functional as F
 from torch.optim import Adam
 from utils import helpers as utl
-from torchkit.constant import *
+from policies.rl.constant import *
 from torchkit.networks import FlattenMlp
-from torchkit.actor import DeterministicPolicy, TanhGaussianPolicy
+from policies.rl.actor import DeterministicPolicy, TanhGaussianPolicy
 import torchkit.pytorch_utils as ptu
 from utils import logger
